字典管理，菜单光临

+++ app/model/WMSModel.go
@@ -5,7 +5,7 @@ import (
 )
 
 type WMSModel struct {
-	Id      int64 `json:"id" gorm"primarykey"`
+	Id      int `json:"id" gorm"primarykey"`
 	Created time.Time `json:"created" gorm:"default:null"`
 	Updated time.Time `json:"updated" gorm:"default:null"`
 }
